Netflix Chatbot Testing
I've tried to integrate a change theme button, ai chatbot support and added some js to the code. i have dont it using the cursor. trust me, its not the best idea to use ai code editors, its helpful sometimes but 70% of the time it makes such horrible mistakes which u can never imagine and it takes days to fix it